backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
black==25.1.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
fastapi==0.116.1
flake8==7.3.0
google-generativeai==0.8.3
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
svix==1.70.1
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
wrapt==1.17.2
